Attempting to fix other junit tests that broke.

+++ src/test/java/seng202/team6/model/entities/AirportTest.java
@@ -16,9 +16,9 @@ public class AirportTest {
     @Test
     public void GetDistanceTest() {
         Airport airport1 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                35, -35,  100, 3, '3');
+                (float)35, (float)-35,  100, 3, '3');
         Airport airport2 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                -35, 35,  100, 3, '3');
+                (float)-35, (float)35,  100, 3, '3');
         double result = airport1.GetDistance(airport2);
 
         assertEquals(10642, Math.floor(result), 1e-15);
@@ -30,11 +30,11 @@ public class AirportTest {
     @Test
     public void TestEqualityDirect() {
         Airport airport1 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                0, 0,  100, 3, '3');
+                (float)0, (float)0,  100, 3, '3');
         Airport airport2 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                0, 0,  100, 3, '3');
+                (float)0, (float)0,  100, 3, '3');
         Airport airport3 = new Airport(1, "WRONG", "VOID", "VOID", "VOID", "VOID",
-                0, 0,  100, 3, '3');
+                (float)0, (float)0,  100, 3, '3');
 
         assertEquals(airport1, airport2);
         assertNotEquals(airport1, airport3);
@@ -46,12 +46,12 @@ public class AirportTest {
     @Test
     public void TestEqualityChain() {
         Airport airport1 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                0, 0,  100, 3, '3');
+                (float)0, (float)0,  100, 3, '3');
         Airport airport2 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                0, 0,  100, 3, '3');
+                (float)0, (float)0,  100, 3, '3');
 
         Airport airport3 = new Airport(1, "CHC", "VOID", "VOID", "VOID", "VOID",
-                0, 0,  100, 3, '3');
+                (float)0, (float)0,  100, 3, '3');
 
         assertEquals(airport1, airport2);
         assertEquals(airport2, airport3);
